+++ api/src/routes/index.js
@@ -4,7 +4,7 @@ const { Router } = require("express");
 const techUsers = require("./techUsers");
 const finalUsers = require("./finalUsers");
 const request = require("./workOrders");
-const auth = require("./auth");
+const login = require("./login");
 
 const router = Router();
 
@@ -14,6 +14,6 @@ const router = Router();
 router.use("/techUsers", techUsers);
 router.use("/finalUsers", finalUsers);
 router.use("/request", request);
-router.use("/auth", auth);
+router.use("/login", login);
 
 module.exports = router;
