+++ api/src/routes/login.js
@@ -1,0 +1,8 @@
+const { Router } = require("express");
+const router = Router();
+
+const login = require("../controllers/Login");
+
+router.post("/", login);
+
+module.exports = router;
